backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.38.0
qwen-agent==0.0.31
python-dotenv==1.0.1
duckduckgo-search==8.1.1
scikit-learn==1.7.2
requests==2.32.3
mcp==1.19.0
